Use tokenService to generate jwt tokens
