+++ backend/user-service/controller/auth-controller.js
@@ -3,6 +3,7 @@ import jwt from "jsonwebtoken";
 import { findUserByEmail as _findUserByEmail, findUserById as _findUserById } from "../model/repository.js";
 import { formatUserResponse } from "./user-controller.js";
 import { jwtConfig, REFRESH_TOKEN_COOKIE_KEY, refreshTokenCookieOptions } from "../config/authConfig.js";
+import { generateAccessToken, generateRefreshToken } from "../service/tokenService.js";
 
 export async function handleLogin(req, res) {
   const { email, password } = req.body;
@@ -19,12 +20,8 @@ export async function handleLogin(req, res) {
       }
 
       // Generate access and refresh token
-      const accessToken = jwt.sign(
-        { id: user.id, isAdmin: user.isAdmin },
-        jwtConfig.accessTokenSecret,
-        jwtConfig.accessTokenOptions
-      );
-      const refreshToken = jwt.sign({ id: user.id }, jwtConfig.refreshTokenSecret, jwtConfig.refreshTokenOptions);
+      const accessToken = generateAccessToken(user);
+      const refreshToken = generateRefreshToken(user);
 
       res.cookie(REFRESH_TOKEN_COOKIE_KEY, refreshToken, refreshTokenCookieOptions);
 
@@ -73,11 +70,7 @@ export async function refresh(req, res) {
     if (!dbUser) {
       return res.status(401).json({ message: "Unauthorized: User not found" });
     }
-    const accessToken = jwt.sign(
-      { id: user.id, isAdmin: user.isAdmin },
-      jwtConfig.accessTokenSecret,
-      jwtConfig.accessTokenOptions
-    );
+    const accessToken = generateAccessToken(user);
     return res.status(200).json({
       message: "Access token refreshed",
       data: accessToken,
