+++ backend/user-service/service/tokenService.js
@@ -1,0 +1,15 @@
+import jwt from "jsonwebtoken";
+import { jwtConfig } from "../config/authConfig";
+
+export const generateAccessToken = (user) => {
+  const accessToken = jwt.sign(
+    { id: user.id, isAdmin: user.isAdmin },
+    jwtConfig.accessTokenSecret,
+    jwtConfig.accessTokenOptions
+  );
+  return accessToken;
+};
+export const generateRefreshToken = (user) => {
+  const refreshToken = jwt.sign({ id: user.id }, jwtConfig.refreshTokenSecret, jwtConfig.refreshTokenOptions);
+  return refreshToken;
+};
